all enums and static const variables are now UPPERCASE (mainly to be stylistically similar to python)

+++ mandala/basic_gpu_program.cpp
@@ -48,8 +48,8 @@ void main()
 		gpu.enable_vertex_attribute_array(location_location);
 		gpu.enable_vertex_attribute_array(color_location);
 
-		gpu.set_vertex_attrib_pointer(location_location, 3, gpu_data_type_e::float_, false, sizeof(vertex_type), reinterpret_cast<void*>(offsetof(vertex_type, location)));
-		gpu.set_vertex_attrib_pointer(color_location, 4, gpu_data_type_e::float_, false, sizeof(vertex_type), reinterpret_cast<void*>(offsetof(vertex_type, color)));
+		gpu.set_vertex_attrib_pointer(location_location, 3, gpu_data_type_e::FLOAT, false, sizeof(vertex_type), reinterpret_cast<void*>(offsetof(vertex_type, location)));
+		gpu.set_vertex_attrib_pointer(color_location, 4, gpu_data_type_e::FLOAT, false, sizeof(vertex_type), reinterpret_cast<void*>(offsetof(vertex_type, color)));
 	}
 
 	void basic_gpu_program_t::on_unbind()
